Extract ipfs.PrepareIPFSConfig() and schema.IPFSCachingRouterDefaultURI
