+++ cmd/start.go
@@ -5,7 +5,6 @@ import (
 	"crypto/rand"
 	"errors"
 	"fmt"
-	"gx/ipfs/QmZNkThpqfVXs9GNbexPrfBbXSLNYeKrE7jwFM2oqHbyqN/go-libp2p-protocol"
 	"io"
 	"io/ioutil"
 	"net"
@@ -25,13 +24,9 @@ import (
 	dhtopts "gx/ipfs/QmSY3nkMNLzh9GdbFKK5tT7YMfLpf52iUZ8ZRkr29MJaa5/go-libp2p-kad-dht/opts"
 	ma "gx/ipfs/QmTZBfrPJmjWsCvHEtX5FE6KimVJhsJg5sBbqEFYf4UZtL/go-multiaddr"
 	config "gx/ipfs/QmUAuYuiafnJRZxDDX7MuruMNsicYNuyub5vUeAcupUBNs/go-ipfs-config"
-	ds "gx/ipfs/QmUadX5EcvrBmxAV9sE7wUWtWSqxns5K84qKJBixmcT1w9/go-datastore"
 	ipnspb "gx/ipfs/QmUwMnKKjH3JwGKNVZ3TcP37W93xzqNA4ECFFiMo6sXkkc/go-ipns/pb"
 	peer "gx/ipfs/QmYVXrKrKHDC9FobgmcmshCDyWwdrfwfanNQN4oxJ9Fk3h/go-libp2p-peer"
-	p2phost "gx/ipfs/QmYrWiWM4qtrnCeT3R14jY3ZZyirDNJgwK57q4qFYePgbd/go-libp2p-host"
 	oniontp "gx/ipfs/QmYv2MbwHn7qcvAPFisZ94w85crQVpwUuv8G7TuUeBnfPb/go-onion-transport"
-	routing "gx/ipfs/QmYxUdYY9S6yg5tSPVin5GFTvtfsLauVcr7reHDD3dM8xf/go-libp2p-routing"
-	record "gx/ipfs/QmbeHtaBy9nZsW4cHRcvgVY4CnDhXudE2Dr6qDxS7yg9rX/go-libp2p-record"
 	ipfslogging "gx/ipfs/QmbkT7eMTyXfpeyB3ZMxxcxg7XH8t6uXp49jqzz4HB7BGF/go-log/writer"
 	manet "gx/ipfs/Qmc85NSvmSG4Frn9Vb2cBc1rMyULH6D3TNVEfCzSKoUpip/go-multiaddr-net"
 	bitswap "gx/ipfs/QmcSPuzpSbVLU6UHU4e5PwZpm4fHbCn5SbNR5ZNL6Mj63G/go-bitswap/network"
@@ -106,13 +101,6 @@ type Start struct {
 }
 
 func (x *Start) Execute(args []string) error {
-
-	if x.Testnet || x.Regtest {
-		ipfscore.DHTOption = constructTestnetDHTRouting
-	} else {
-		ipfscore.DHTOption = constructDHTRouting
-	}
-
 	printSplashScreen(x.Verbose)
 
 	if x.Testnet && x.Regtest {
@@ -274,7 +262,6 @@ func (x *Start) Execute(args []string) error {
 		log.Error("scan ipns extra config:", err)
 		return err
 	}
-	apiRouterURI = ipnsExtraConfig.APIRouter
 
 	// IPFS node setup
 	r, err := fsrepo.Open(repoPath)
@@ -397,24 +384,7 @@ func (x *Start) Execute(args []string) error {
 		cfg.Swarm.DisableNatPortMap = true
 	}
 
-	ncfg := &ipfscore.BuildCfg{
-		Repo:   r,
-		Online: true,
-		ExtraOpts: map[string]bool{
-			"mplex":  true,
-			"ipnsps": true,
-		},
-	}
-	if x.Regtest {
-		ncfg.Routing = constructDHTRouting
-	} else {
-		if x.Testnet {
-			ncfg.Routing = constructTestnetDHTRouting
-		} else {
-			ncfg.Routing = constructRouting
-		}
-	}
-
+	ncfg := ipfs.PrepareIPFSConfig(r, ipnsExtraConfig.APIRouter, x.Testnet, x.Regtest)
 	nd, err := ipfscore.NewNode(cctx, ncfg)
 	if err != nil {
 		log.Error("create new ipfs node:", err)
@@ -864,45 +834,6 @@ func newHTTPGateway(node *core.OpenBazaarNode, ctx commands.Context, authCookie 
 	ml := logging.MultiLogger(apiFileFormatter)
 
 	return api.NewGateway(node, authCookie, manet.NetListener(gwLis), config, ml, opts...)
-}
-
-func constructRouting(ctx context.Context, host p2phost.Host, dstore ds.Batching, validator record.Validator) (routing.IpfsRouting, error) {
-	dhtRouting, err := dht.New(
-		ctx, host,
-		dhtopts.Datastore(dstore),
-		dhtopts.Validator(validator),
-	)
-	if err != nil {
-		return nil, err
-	}
-	apiRouter := ipfs.NewAPIRouter(apiRouterURI)
-	cachingRouter := ipfs.NewCachingRouter(dhtRouting, &apiRouter)
-	return cachingRouter, nil
-}
-
-func constructDHTRouting(ctx context.Context, host p2phost.Host, dstore ds.Batching, validator record.Validator) (routing.IpfsRouting, error) {
-	return dht.New(
-		ctx, host,
-		dhtopts.Datastore(dstore),
-		dhtopts.Validator(validator),
-	)
-}
-
-func constructTestnetDHTRouting(ctx context.Context, host p2phost.Host, dstore ds.Batching, validator record.Validator) (routing.IpfsRouting, error) {
-	testnetDHT := protocol.ID("/openbazaar/kad/testnet/1.0.0")
-	testnetApp := protocol.ID("/openbazaar/app/testnet/1.0.0")
-	dhtRouting, err := dht.New(
-		ctx, host,
-		dhtopts.Datastore(dstore),
-		dhtopts.Validator(validator),
-		dhtopts.Protocols(testnetDHT, testnetApp),
-	)
-	if err != nil {
-		return nil, err
-	}
-	apiRouter := ipfs.NewAPIRouter(apiRouterURI)
-	cachingRouter := ipfs.NewCachingRouter(dhtRouting, &apiRouter)
-	return cachingRouter, nil
 }
 
 // serveHTTPApi collects options, creates listener, prints status message and starts serving requests
